Fix compilation for XCode 14.3
The default argument of `type::create(...)` was `std::vector<argument>()`. When it is defined, `argument` is still an incomplete type. Previous compiler versions seemed to be OK with this by deferring the constructor call.

XCode 14.3, however, demands that `argument` be fully defined when `type::create(...)` is declared.

This change replaces the default argument with a separate `type::create()` declaration whose implementation simply calls `type::create(...)` with an empty vector.